Minor improvements
Using throw instead of Write-Warning for ending failures.

Prevent looping forever in OpenVPN by having max retries before aborting

+++ Windows/Prepare-PC/resources/installers/1-OpenVPN/install.ps1
@@ -18,6 +18,7 @@
 # Install OpenVPN along with the config file
 $envTEMP = (Get-Item -LiteralPath $env:TEMP).FullName # Required due to PowerShell bug with shortnames appearing when they shouldn't be
 $loopDelay = 1 # second
+$maxTries = 10 # times before aborting failed attempt to find download
 $appTitle = 'OpenVPN'
 $configOVPN = "${PSScriptRoot}\example-config.ovpn"
 $ovpnFolder = $env:ProgramFiles + "\OpenVPN"
@@ -34,38 +35,46 @@ if ($ovpnWasPreinstalled) {
   Write-Output "Downloading ${appTitle}..."
   Write-Output '' # Makes log look better
   # need to download the installer first
-  $ovpnDownloadURL = $null
-  while (-Not $ovpnDownloadURL) {
+  $ovpnDownloadURL = $Null
+  $ovpnDownloadAttempt = 0
+  while (($ovpnDownloadAttempt -lt $maxTries) -And (-Not $ovpnDownloadURL)) {
     # need to loop until an MSI download URL is pulled from downloads
     $ovpnDownloadsGET = Invoke-WebRequest -Uri $ovpnDownloadsURL -UseBasicParsing
     if ($ovpnDownloadsGET.StatusCode -eq 200) {
       $ovpnDownloadURL = [Regex]::Matches($ovpnDownloadsGET.Content, $ovpnDownloadPattern)[0].Value
-    } else { Start-Sleep -Seconds $loopDelay }
-  }
-  $tempOvpnMSI = $envTEMP + '\' + $ovpnDownloadURL.substring($ovpnDownloadURL.LastIndexOf('/') + 1)
-  while ((Invoke-WebRequest -Uri $ovpnDownloadURL -OutFile $tempOvpnMSI -UseBasicParsing -PassThru).StatusCode -ne 200) {
-    # need to loop until OpenVPN installer is downloaded
-    Start-Sleep -Seconds $loopDelay
-  }
-  Write-Output "Downloaded ${appTitle}."
-  Write-Output '' # Makes log look better
-  Write-Output "Installing ${appTitle}..."
-  Write-Output '' # Makes log look better
-  New-Item -ItemType Directory -Force -Path $ovpnConfigFolder -ErrorAction SilentlyContinue | Out-Null
-  $configCopied = Copy-Item -Path $configOVPN -Destination $ovpnConfigFolder -PassThru -Force
-  $setupArgs = "/i `"${tempOvpnMSI}`" /qn"
-  $ovpnInstall = Start-Process 'msiexec.exe' -ArgumentList $setupArgs -NoNewWindow -PassThru -Wait
-  Remove-Item -Path $tempOvpnMSI -Force -ErrorAction SilentlyContinue
-  if ($configCopied -And (0 -eq $ovpnInstall.ExitCode)) {
-    Write-Output "Successfully installed ${appTitle}."
-  } else {
-    $ovpnInstallError = ""
-    if ($configCopied) {
-      $ovpnInstallError += "the config being unable to copy"
-      if ($ovpnInstall) { $ovpnInstallError += ' ' }
+    } else {
+      $ovpnDownloadAttempt++
+      Start-Sleep -Seconds $loopDelay
     }
-    if (0 -ne $ovpnInstall.ExitCode) { $ovpnInstallError += "the installer failing (exit code: $($ovpnInstall.ExitCode))" }
-    Write-Warning "Failed to install ${appTitle}, due to ${ovpnInstallError}."
+  }
+  if ($ovpnDownloadURL) {
+    $tempOvpnMSI = $envTEMP + '\' + $ovpnDownloadURL.substring($ovpnDownloadURL.LastIndexOf('/') + 1)
+    while ((Invoke-WebRequest -Uri $ovpnDownloadURL -OutFile $tempOvpnMSI -UseBasicParsing -PassThru).StatusCode -ne 200) {
+      # need to loop until OpenVPN installer is downloaded
+      Start-Sleep -Seconds $loopDelay
+    }
+    Write-Output "Downloaded ${appTitle}."
+    Write-Output '' # Makes log look better
+    Write-Output "Installing ${appTitle}..."
+    Write-Output '' # Makes log look better
+    New-Item -ItemType Directory -Force -Path $ovpnConfigFolder -ErrorAction SilentlyContinue | Out-Null
+    $configCopied = Copy-Item -Path $configOVPN -Destination $ovpnConfigFolder -PassThru -Force
+    $setupArgs = "/i `"${tempOvpnMSI}`" /qn"
+    $ovpnInstall = Start-Process 'msiexec.exe' -ArgumentList $setupArgs -NoNewWindow -PassThru -Wait
+    Remove-Item -Path $tempOvpnMSI -Force -ErrorAction SilentlyContinue
+    if ($configCopied -And (0 -eq $ovpnInstall.ExitCode)) {
+      Write-Output "Successfully installed ${appTitle}."
+    } else {
+      $ovpnInstallError = ""
+      if ($configCopied) {
+        $ovpnInstallError += "the config being unable to copy"
+        if ($ovpnInstall) { $ovpnInstallError += ' ' }
+      }
+      if (0 -ne $ovpnInstall.ExitCode) { $ovpnInstallError += "the installer failing (exit code: $($ovpnInstall.ExitCode))" }
+      throw "Failed to install ${appTitle}, due to ${ovpnInstallError}."
+    }
+  } else {
+    throw "Failed to download installer for ${appTitle}, the API or website might be down."
   }
 }
 Write-Output '' # Makes log look better
